feat: refactor transaction history page to use TransactionCard component and remove inline card implementation
